Update run-tests-on-maelstrom-inner.
Rewrite and simplify it. Change it to use maelstrom-admin to stop the broker.

+++ scripts/run-tests-on-maelstrom-inner.sh
@@ -15,25 +15,22 @@ if [[ $# -gt 0 ]]; then
     export MAELSTROM_GITHUB_ACTIONS_TOKEN=$1
     export MAELSTROM_GITHUB_ACTIONS_URL=$2
     export MAELSTROM_CLUSTER_COMMUNICATION_STRATEGY=github
-    START_WORKER=0
     echo "using github for artifact transfer"
-fi
-
-set +e
-
-TEMPFILE=$(mktemp --tmpdir run-tests-on-maelstrom-broker-stderr.XXXXXX)
-cargo run --release --bin maelstrom-broker -- $BROKER_ARGS 2> >(tee "$TEMPFILE" >&2) &
-BROKER_PID=$!
-PORT=$( \
-	tail -f "$TEMPFILE" \
-	| awk '/\<addr: / { print $0; exit}' \
-	| sed -Ee 's/^.*\baddr: [^,]*:([0-9]+),.*$/\1/' \
-)
-if [[ $START_WORKER -gt 0 ]]; then
+    cargo run --release --bin maelstrom-broker -- $BROKER_ARGS &
+else 
+    TEMPFILE=$(mktemp --tmpdir run-tests-on-maelstrom-broker-stderr.XXXXXX)
+    cargo run --release --bin maelstrom-broker -- $BROKER_ARGS 2> >(tee "$TEMPFILE" >&2) &
+    PORT=$( \
+    	tail -f "$TEMPFILE" \
+    	| awk '/\<addr: / { print $0; exit}' \
+    	| sed -Ee 's/^.*\baddr: [^,]*:([0-9]+),.*$/\1/' \
+    )
     cargo run --release --bin maelstrom-worker -- --broker=localhost:$PORT $WORKER_ARGS &
     CARGO_ARGS="$CARGO_ARGS --broker=localhost:$PORT"
     PYTEST_ARGS="$PYTEST_ARGS --broker=localhost:$PORT"
 fi
+
+set +e
 
 cargo run --release --bin cargo-maelstrom -- $CARGO_ARGS
 CARGO_MAELSTROM_STATUS=$?
@@ -41,16 +38,20 @@ CARGO_MAELSTROM_STATUS=$?
 cargo run --release --bin maelstrom-pytest -- $PYTEST_ARGS
 MAELSTROM_PYTEST_STATUS=$?
 
-kill -15 $BROKER_PID
-rm "$TEMPFILE"
-set -e
+cargo run --release --bin maelstrom-admin stop
 
-if [ $CARGO_MAELSTROM_STATUS != 0 ]; then
-	exit $CARGO_MAELSTROM_STATUS
+if [[ -v TEMPFILE ]]; then
+    rm "$TEMPFILE"
 fi
 
-if [ $MAELSTROM_PYTEST_STATUS != 0 ]; then
-	exit $MAELSTROM_PYTEST_STATUS
+set -e
+
+if [[ $CARGO_MAELSTROM_STATUS != 0 ]]; then
+    exit $CARGO_MAELSTROM_STATUS
+fi
+
+if [[ $MAELSTROM_PYTEST_STATUS != 0 ]]; then
+    exit $MAELSTROM_PYTEST_STATUS
 fi
 
 cargo test --release --doc
